fixup! Add pause test
